+++ src/constants/chat.ts
@@ -1,3 +1,4 @@
 export enum CHAT {
+  ROOM = "chat-room",
   LIST = "chat-list",
 }
