+++ src/lib/tanstack/quries/index.ts
@@ -1,9 +1,21 @@
 import type { QueryKey } from "@tanstack/react-query";
-import { useSuspenseQuery } from "@tanstack/react-query";
+import { QueryClient, useSuspenseQuery } from "@tanstack/react-query";
 
 // import { api } from "@/src/apis";
 import { mockApi } from "@/src/apis/mock";
 import type { BaseResponseDTO } from "@/src/models";
+
+export const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      staleTime: 60 * 1000,
+    },
+  },
+});
+
+export const invalidateQueries = (queryKey: QueryKey) => {
+  queryClient.invalidateQueries({ queryKey });
+};
 
 export const useCommonSuspenseQuery = <T>(prefix: string, queryKey: QueryKey, url: string) => {
   return useSuspenseQuery<BaseResponseDTO<T>>({
